requirements.txt
certifi==2025.8.3
cffi==2.0.0
et_xmlfile==2.0.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
